fix(lens): move the fleet dial-state/snapshot vocabulary into a no-transport leaf
FleetHostStatus, FleetHostState, and FleetSnapshot now live in a new fleetTypes.ts leaf that both fleet.ts (their producer) and render.ts (their consumer) import, so the dependency arrow points OUT and the orchestrator no longer reaches back into the renderer for the type that names its own output. FleetUrgency/FleetGroup/FleetView/FleetSummary stay in render.ts as genuine projection types.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-2, raised by hickey). Not pushed or merged.
